test(rust): clarify retained callback contracts
