+++ tests/test_litellm_rust/chat/test_callback_mutation.py
@@ -44,62 +44,136 @@ def _verify_sigv4(request: RecordedRequest, secret_key: str) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("asynchronous", [False, True])
-@pytest.mark.parametrize("provider", ["anthropic", "bedrock"])
-@pytest.mark.parametrize("rebind_logging_view", [False, True])
-@pytest.mark.parametrize("status", [200, 429])
-@pytest.mark.parametrize("native", [False, True])
-async def test_chat_retains_callback_edits_through_public_dispatch(
+@pytest.mark.parametrize("asynchronous", [False, True], ids=["sync", "async"])
+@pytest.mark.parametrize("rebind_logging_view", [False, True], ids=["in-place-view", "rebound-view"])
+@pytest.mark.parametrize("native", [False, True], ids=["python", "rust"])
+async def test_anthropic_pre_call_body_and_header_mutations_reach_provider(
     recording_server: RecordingServer,
     asynchronous: bool,
-    provider: str,
     rebind_logging_view: bool,
-    status: int,
     native: bool,
 ) -> None:
-    import threading
-
     litellm.rust(native)
 
     from tests.test_litellm_rust.support.callback_recorder import RecordingLogger
     from tests.test_litellm_rust.support.requests import MESSAGES_RESPONSE
 
-    recording_server.default_response = ResponseSpec(
-        status=status,
-        body=(
-            MESSAGES_RESPONSE
-            if provider == "anthropic"
-            else {
-                "output": {"message": {"role": "assistant", "content": [{"text": "native chat"}]}},
-                "stopReason": "end_turn",
-                "usage": {"inputTokens": 5, "outputTokens": 4, "totalTokens": 9},
-                "metrics": {"latencyMs": 1},
-            }
-        ),
-    )
-    caller_thread: Final = threading.current_thread()
-    observations: Final = []
+    recording_server.default_response = ResponseSpec(body=MESSAGES_RESPONSE)
+    callback_received_dict: Final = []
 
     class EditingLogger(RecordingLogger):
         def log_pre_api_call(self, model, messages, kwargs):
             super().log_pre_api_call(model, messages, kwargs)
             body = kwargs["additional_args"]["complete_input_dict"]
             headers = kwargs["additional_args"]["headers"]
-            if provider == "anthropic":
-                assert isinstance(body, dict)
-                body["messages"][0]["content"][0]["text"] = "edited by callback"
-                body["max_tokens"] = 32
-            else:
-                assert isinstance(body, str)
-                assert json.loads(body)["messages"][0]["content"][0]["text"] == "original"
+            callback_received_dict.append(isinstance(body, dict))
+            if not isinstance(body, dict):
+                return
+            body["messages"][0]["content"][0]["text"] = "edited by callback"
+            body["max_tokens"] = 32
             headers["x-retained-callback"] = "original"
-            observations.append((threading.current_thread(), body, headers))
             if rebind_logging_view:
                 kwargs["additional_args"]["complete_input_dict"] = {"replacement": True}
                 kwargs["additional_args"]["headers"] = {"x-retained-callback": "replacement"}
 
     recorder: Final = EditingLogger()
     kwargs: Final = {
+        "model": "anthropic/claude-opus-5",
+        "messages": [{"role": "user", "content": "original"}],
+        "max_tokens": 64,
+        "api_key": "test-key",
+        "api_base": recording_server.base_url,
+        "callbacks": [recorder],
+        "num_retries": 0,
+    }
+    response: Final = await litellm.acompletion(**kwargs) if asynchronous else litellm.completion(**kwargs)
+
+    assert response._hidden_params.get("additional_headers", {}).get("x-litellm-rust") == ("true" if native else None)
+    assert callback_received_dict == [True]
+    assert recorder.names.count("log_pre_api_call") == 1
+    assert recording_server.requests[0].body["messages"][0]["content"][0]["text"] == "edited by callback"
+    assert recording_server.requests[0].headers["x-retained-callback"] == "original"
+    assert recording_server.requests[0].body["max_tokens"] == 32
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize("asynchronous", [False, True], ids=["sync", "async"])
+@pytest.mark.parametrize("rebind_logging_view", [False, True], ids=["in-place-view", "rebound-view"])
+@pytest.mark.parametrize("native", [False, True], ids=["python", "rust"])
+async def test_bedrock_pre_call_rebinding_does_not_replace_signed_transport(
+    recording_server: RecordingServer,
+    asynchronous: bool,
+    rebind_logging_view: bool,
+    native: bool,
+) -> None:
+    from tests.test_litellm_rust.support.callback_recorder import RecordingLogger
+
+    litellm.rust(native)
+    recording_server.default_response = ResponseSpec(
+        body={
+            "output": {"message": {"role": "assistant", "content": [{"text": "native chat"}]}},
+            "stopReason": "end_turn",
+            "usage": {"inputTokens": 5, "outputTokens": 4, "totalTokens": 9},
+            "metrics": {"latencyMs": 1},
+        }
+    )
+    callback_body_text: Final = []
+
+    class EditingLogger(RecordingLogger):
+        def log_pre_api_call(self, model, messages, kwargs):
+            super().log_pre_api_call(model, messages, kwargs)
+            additional: Final = kwargs["additional_args"]
+            body: Final = additional["complete_input_dict"]
+            callback_body_text.append(
+                json.loads(body)["messages"][0]["content"][0]["text"] if isinstance(body, str) else None
+            )
+            additional["headers"]["x-retained-callback"] = "original"
+            if rebind_logging_view:
+                additional["complete_input_dict"] = {"replacement": True}
+                additional["headers"] = {"x-retained-callback": "replacement"}
+
+    recorder: Final = EditingLogger()
+    request: Final = {
+        "model": "bedrock/anthropic.claude-opus-5",
+        "messages": [{"role": "user", "content": "original"}],
+        "max_tokens": 64,
+        "api_base": recording_server.base_url,
+        "callbacks": [recorder],
+        "num_retries": 0,
+        "aws_access_key_id": "test-access",
+        "aws_secret_access_key": "test-secret",
+        "aws_region_name": "us-east-1",
+    }
+    response: Final = await litellm.acompletion(**request) if asynchronous else litellm.completion(**request)
+    recorded: Final = recording_server.requests[0]
+
+    assert response._hidden_params.get("additional_headers", {}).get("x-litellm-rust") == ("true" if native else None)
+    assert callback_body_text == ["original"]
+    assert recorder.names.count("log_pre_api_call") == 1
+    assert recorded.body["messages"][0]["content"][0]["text"] == "original"
+    assert "replacement" not in recorded.body
+    assert recorded.headers["x-retained-callback"] == "original"
+    assert recorded.body["inferenceConfig"]["maxTokens"] == 64
+    assert recorded.headers["authorization"].startswith("AWS4-HMAC-SHA256 ")
+    _verify_sigv4(recorded, "test-secret")
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize("asynchronous", [False, True], ids=["sync", "async"])
+@pytest.mark.parametrize("provider", ["anthropic", "bedrock"])
+@pytest.mark.parametrize("native", [False, True], ids=["python", "rust"])
+async def test_chat_provider_failure_reaches_one_failure_callback(
+    recording_server: RecordingServer,
+    asynchronous: bool,
+    provider: str,
+    native: bool,
+) -> None:
+    from tests.test_litellm_rust.support.callback_recorder import RecordingLogger
+
+    litellm.rust(native)
+    recording_server.default_response = ResponseSpec(status=429, body={"message": "rate limited"})
+    recorder: Final = RecordingLogger()
+    request: Final = {
         "model": "anthropic/claude-opus-5" if provider == "anthropic" else "bedrock/anthropic.claude-opus-5",
         "messages": [{"role": "user", "content": "original"}],
         "max_tokens": 64,
@@ -113,45 +187,23 @@ async def test_chat_retains_callback_edits_through_public_dispatch(
             else {}
         ),
     }
-    if status != 200:
-        with pytest.raises((litellm.APIError, litellm.RateLimitError)) as raised:
-            await litellm.acompletion(**kwargs) if asynchronous else litellm.completion(**kwargs)
-        failure_event: Final = "async_log_failure_event" if asynchronous else "log_failure_event"
-        failures: Final = await recorder.wait_for_async(failure_event)
-        assert raised.value.status_code == status
-        assert failures[0].kwargs["exception"] is raised.value
-        assert recorder.names.count(failure_event) == 1
-        assert recorder.names.count("log_pre_api_call") == 1
-        assert len(recording_server.requests) == 1
-        return
 
-    response: Final = await litellm.acompletion(**kwargs) if asynchronous else litellm.completion(**kwargs)
-    event_name: Final = "async_log_success_event" if asynchronous else "log_success_event"
-    events: Final = await recorder.wait_for_async(event_name)
+    with pytest.raises((litellm.APIError, litellm.RateLimitError)) as raised:
+        await litellm.acompletion(**request) if asynchronous else litellm.completion(**request)
+    failure_event: Final = "async_log_failure_event" if asynchronous else "log_failure_event"
+    failures: Final = await recorder.wait_for_async(failure_event)
 
-    if native:
-        assert response._hidden_params["additional_headers"]["x-litellm-rust"] == "true"
+    assert raised.value.status_code == 429
+    assert failures[0].kwargs["exception"] is raised.value
+    assert recorder.names.count(failure_event) == 1
     assert recorder.names.count("log_pre_api_call") == 1
-    assert recorder.names.count(event_name) == 1
-    if asynchronous and provider == "anthropic" and not native:
-        assert observations[0][0] is not caller_thread
-    else:
-        assert observations[0][0] is caller_thread
-    assert events[0].response is response
-    assert recording_server.requests[0].body["messages"][0]["content"][0]["text"] == (
-        "edited by callback" if provider == "anthropic" else "original"
-    )
-    assert recording_server.requests[0].headers["x-retained-callback"] == "original"
-    body: Final = recording_server.requests[0].body
-    assert (body["max_tokens"] if provider == "anthropic" else body["inferenceConfig"]["maxTokens"]) == (
-        32 if provider == "anthropic" else 64
-    )
+    assert len(recording_server.requests) == 1
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("asynchronous", [False, True])
+@pytest.mark.parametrize("asynchronous", [False, True], ids=["sync", "async"])
 @pytest.mark.parametrize("status", [200, 429])
-@pytest.mark.parametrize("native", [False, True])
+@pytest.mark.parametrize("native", [False, True], ids=["python", "rust"])
 async def test_bedrock_callbacks_share_state_without_replacing_signed_transport(
     recording_server: RecordingServer,
     monkeypatch: pytest.MonkeyPatch,
@@ -173,6 +225,7 @@ async def test_bedrock_callbacks_share_state_without_replacing_signed_transport(
         },
     )
     observations: Final[list[tuple[int, int]]] = []
+    shared_state: Final[list[tuple[object, object, object]]] = []
     replacement_body: Final = {"replacement": True}
     rebound_headers: Final = {"x-callback-header": "rebound"}
 
@@ -182,8 +235,15 @@ async def test_bedrock_callbacks_share_state_without_replacing_signed_transport(
             callback_body: Final = additional["complete_input_dict"]
             original_headers: Final = additional["headers"]
             observations.append((id(kwargs), id(additional)))
-            assert isinstance(callback_body, str)
-            assert json.loads(callback_body)["messages"][0]["content"][0]["text"] == "original"
+            shared_state.append(
+                (
+                    isinstance(callback_body, str),
+                    json.loads(callback_body)["messages"][0]["content"][0]["text"]
+                    if isinstance(callback_body, str)
+                    else None,
+                    None,
+                )
+            )
             kwargs["callback_marker"] = "visible"
             original_headers["x-callback-header"] = "in-place"
             additional["complete_input_dict"] = replacement_body
@@ -194,9 +254,13 @@ async def test_bedrock_callbacks_share_state_without_replacing_signed_transport(
         def log_pre_api_call(self, model, messages, kwargs):
             additional: Final = kwargs["additional_args"]
             observations.append((id(kwargs), id(additional)))
-            assert kwargs["callback_marker"] == "visible"
-            assert additional["complete_input_dict"] is replacement_body
-            assert additional["headers"] is rebound_headers
+            shared_state.append(
+                (
+                    kwargs.get("callback_marker"),
+                    additional["complete_input_dict"] is replacement_body,
+                    additional["headers"] is rebound_headers,
+                )
+            )
 
     request: Final = {
         "model": "bedrock/anthropic.claude-opus-5",
@@ -216,6 +280,7 @@ async def test_bedrock_callbacks_share_state_without_replacing_signed_transport(
             await litellm.acompletion(**request) if asynchronous else litellm.completion(**request)
 
     assert observations[0] == observations[1]
+    assert shared_state == [(True, "original", None), ("visible", True, True)]
     recorded: Final = recording_server.requests[0]
     assert recorded.body["messages"][0]["content"][0]["text"] == "original"
     assert "replacement" not in recorded.body
@@ -261,7 +326,7 @@ async def test_native_chat_pre_call_preserves_the_caller_task_and_context(
 
 
 @pytest.mark.asyncio
-async def test_native_chat_concurrent_calls_keep_callback_roots_isolated(
+async def test_chat_concurrent_calls_keep_callback_roots_isolated(
     recording_server: RecordingServer,
 ) -> None:
     import asyncio
@@ -291,8 +356,8 @@ async def test_native_chat_concurrent_calls_keep_callback_roots_isolated(
 
     logger: Final = Correlate()
 
-    async def invoke(call_id: str) -> None:
-        await litellm.acompletion(
+    async def invoke(call_id: str) -> object:
+        return await litellm.acompletion(
             model="anthropic/claude-opus-5",
             messages=[{"role": "user", "content": f"original-{call_id}"}],
             max_tokens=16,
@@ -308,13 +373,14 @@ async def test_native_chat_concurrent_calls_keep_callback_roots_isolated(
         assert all(await asyncio.gather(*(asyncio.to_thread(signal.wait, 3) for signal in accepted)))
     finally:
         release.set()
-    await asyncio.gather(*tasks)
+    responses: Final = await asyncio.gather(*tasks)
     await drain_logging()
 
     sent: Final = {request.body["messages"][0]["content"][0]["text"] for request in recording_server.requests}
     assert sent == {f"callback-{call_id}" for call_id in call_ids}
     assert len(terminal_state) == len(call_ids)
     assert all(token is tokens[call_id] for call_id, token in terminal_state)
+    assert all(response._hidden_params["additional_headers"]["x-litellm-rust"] == "true" for response in responses)
 
 
 @pytest.mark.asyncio
@@ -364,7 +430,7 @@ async def test_native_chat_callback_can_make_a_nested_native_call(
 
 
 @pytest.mark.asyncio
-async def test_native_chat_cancellation_during_io_does_not_publish_a_terminal(
+async def test_chat_cancellation_during_io_does_not_publish_a_terminal(
     recording_server: RecordingServer,
 ) -> None:
     import asyncio
